поменял int на decimal

+++ ControlEsche/Classes/Clothing.cs
@@ -2,7 +2,7 @@ namespace Lab9
 {
     class Clothing : Product
     {
-        public Clothing(string name, int quantity, int price) : base(name,  quantity,  price)
+        public Clothing(string name, int quantity, decimal price) : base(name,  quantity,  price)
         {
             this.name = name;
             this.quantity = quantity;
